# Simplify two queries in channel store