#### Summary
Query in GetMoreChannels was executed with 'dependant subquery' in MySQL <= 5.7. Now it is ~20% faster in MySQL and ~10% faster in PostgreSQL.
Query in getSearchGroupChannelsQuery is around 20% faster in MySQL. In PostgreSQL it is, depending on data, up to a few times faster but sometimes ~20% slower.