@jespino 2019-11-20
@Pomyk Hi, thanks for your PR, I have to think about it and see if we are not changing the behavior here.