@Pomyk 2019-11-20
I think it is equivalent. Upon further inspection it looks like GetMoreChannels function is not used anymore?
I'm also working on a benchmark and from first tests the simplified getSearchGroupChannelsQuery is not faster on MySQL. I will test on PostgreSQL later today.