@Pomyk 2019-11-20
Just ran the benchmark and for MySQL the new version looks a bit faster actually, but for PostgreSQL it's slower. 

    benchmark                                                      old ns/op     new ns/op     delta
    BenchmarkChannelStore/mysql/GetMoreChannels-16                 655245        591167        -9.78%
    BenchmarkChannelStore/mysql/SearchGroupChannels-16             755131        694325        -8.05%
    BenchmarkChannelStore/postgres/GetMoreChannels-16              1143659       1086207       -5.02%
    BenchmarkChannelStore/postgres/SearchGroupChannels-16          1523730       1974434       +29.58%