@Pomyk 2019-11-22
I will close this for now, needs more testing :)